+++ libs/viewer/src/ViewerGui.cpp
@@ -34,7 +34,6 @@
 
 #include <string>
 #include <vector>
-#include <sstream>
 
 using namespace filagui;
 using namespace filament::math;
@@ -394,30 +393,34 @@ ViewerGui::~ViewerGui() {
     delete mImGuiHelper;
 }
 
-void ViewerGui::populateScene(FilamentAsset* asset,  FilamentInstance* instanceToAnimate) {
+void ViewerGui::setAsset(FilamentAsset* asset,  FilamentInstance* instanceToAnimate) {
     if (mAsset != asset) {
         removeAsset();
         mAsset = asset;
+        mVisibleScenes.reset();
+        mVisibleScenes.set(0);
         mCurrentCamera = 0;
         if (!asset) {
             mAnimator = nullptr;
             return;
         }
         mAnimator = instanceToAnimate ? instanceToAnimate->getAnimator() : asset->getAnimator();
+        assert_invariant(mAnimator && "Call loadResources or asyncBeginLoad before setAsset.");
         updateRootTransform();
         mScene->addEntities(asset->getLightEntities(), asset->getLightEntityCount());
+        auto& tcm = mEngine->getRenderableManager();
+        for (size_t i = 0, n = asset->getRenderableEntityCount(); i < n; i++) {
+            auto ri = tcm.getInstance(asset->getRenderableEntities()[i]);
+            tcm.setScreenSpaceContactShadows(ri, true);
+        }
     }
+}
 
-    auto& tcm = mEngine->getRenderableManager();
-
+void ViewerGui::populateScene() {
     static constexpr int kNumAvailable = 128;
     utils::Entity renderables[kNumAvailable];
     while (size_t numWritten = mAsset->popRenderables(renderables, kNumAvailable)) {
-        for (size_t i = 0; i < numWritten; i++) {
-            auto ri = tcm.getInstance(renderables[i]);
-            tcm.setScreenSpaceContactShadows(ri, true);
-        }
-        mScene->addEntities(renderables, numWritten);
+        mAsset->addEntitiesToScene(*mScene, renderables, numWritten, mVisibleScenes);
     }
 }
 
@@ -464,6 +467,34 @@ void ViewerGui::updateIndirectLight() {
     if (mIndirectLight) {
         mIndirectLight->setIntensity(mSettings.lighting.iblIntensity);
         mIndirectLight->setRotation(mat3f::rotation(mSettings.lighting.iblRotation, float3{ 0, 1, 0 }));
+    }
+}
+
+void ViewerGui::sceneSelectionUI() {
+    // Build a list of checkboxes, one for each glTF scene.
+    bool changed = false;
+    for (size_t i = 0, n = mAsset->getSceneCount(); i < n; ++i) {
+        bool isVisible = mVisibleScenes.test(i);
+        const char* name = mAsset->getSceneName(i);
+        if (name) {
+            changed = ImGui::Checkbox(name, &isVisible) || changed;
+        } else {
+            char label[16];
+            snprintf(label, 16, "Scene %zu", i);
+            changed = ImGui::Checkbox(label, &isVisible) || changed;
+        }
+        if (isVisible) {
+            mVisibleScenes.set(i);
+        } else {
+            mVisibleScenes.unset(i);
+        }
+    }
+    // If any checkboxes have been toggled, rebuild the scene list.
+    if (changed) {
+        const utils::Entity* entities = mAsset->getRenderableEntities();
+        const size_t entityCount = mAsset->getRenderableEntityCount();
+        mScene->removeEntities(entities, entityCount);
+        mAsset->addEntitiesToScene(*mScene, entities, entityCount, mVisibleScenes);
     }
 }
 
@@ -754,6 +785,8 @@ void ViewerGui::updateUserInterface() {
             ImGui::SliderFloat("Stride", &ssrefl.stride, 1.0, 10.0f);
         }
         ImGui::Unindent();
+
+        ImGui::Checkbox("Screen-space Guard Band", &mSettings.view.guardBand.enabled);
     }
 
     if (ImGui::CollapsingHeader("Dynamic Resolution")) {
@@ -847,18 +880,24 @@ void ViewerGui::updateUserInterface() {
     if (ImGui::CollapsingHeader("Scene")) {
         ImGui::Indent();
 
-        ImGui::Checkbox("Scale to unit cube", &mSettings.viewer.autoScaleEnabled);
-        updateRootTransform();
+        if (ImGui::Checkbox("Scale to unit cube", &mSettings.viewer.autoScaleEnabled)) {
+            updateRootTransform();
+        }
 
         ImGui::Checkbox("Show skybox", &mSettings.viewer.skyboxEnabled);
         ImGui::ColorEdit3("Background color", &mSettings.viewer.backgroundColor.r);
 
-        // We do not yet support ground shadow in remote mode,
+        // We do not yet support ground shadow or scene selection in remote mode.
         if (!isRemoteMode()) {
             ImGui::Checkbox("Ground shadow", &mSettings.viewer.groundPlaneEnabled);
             ImGui::Indent();
             ImGui::SliderFloat("Strength", &mSettings.viewer.groundShadowStrength, 0.0f, 1.0f);
             ImGui::Unindent();
+
+            if (mAsset->getSceneCount() > 1) {
+                ImGui::Separator();
+                sceneSelectionUI();
+            }
         }
 
         ImGui::Unindent();