+++ samples/gltf_viewer.cpp
@@ -507,6 +507,7 @@ int main(int argc, char** argv) {
         }
 
         loadResources(filename);
+        app.viewer->setAsset(app.asset);
 
         createGroundPlane(engine, scene, app);
 
@@ -665,8 +666,8 @@ int main(int argc, char** argv) {
         // Optionally fit the model into a unit cube at the origin.
         app.viewer->updateRootTransform();
 
-        // Add renderables to the scene as they become ready.
-        app.viewer->populateScene(app.asset);
+        // Gradually add renderables to the scene as their textures become ready.
+        app.viewer->populateScene();
 
         app.viewer->applyAnimation(now);
     };
@@ -766,6 +767,7 @@ int main(int argc, char** argv) {
         app.assetLoader->destroyAsset(app.asset);
         loadAsset(path);
         loadResources(path);
+        app.viewer->setAsset(app.asset);
     });
 
     filamentApp.run(app.config, setup, cleanup, gui, preRender, postRender);